docs: 스토리북 디자인 토큰 문서에서 Semantic 과 Scale Color 분리 (#33)
