+++ packages/ui/src/storybook-components/index.ts
@@ -3,3 +3,6 @@ export { default as PageTitle } from './PageTitle';
 export { default as Section } from './Section';
 export { default as SectionTitle } from './SectionTitle';
 export { default as HowToUse } from './HowToUse';
+export { default as HowToUseClass } from './HowToUseClass';
+export { default as ColorList } from './ColorList';
+export { default as ColorPalette } from './ColorPalette';
